🔥 Refactor: file separation
